Exercise 4.21 complete

+++ controllers/blogs.js
@@ -54,8 +54,17 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.id.toString() === decodedToken.id) {
+        await Blog.findByIdAndDelete(request.params.id);
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'Invalid permissions for delete' })
+    }
 })
 
 module.exports = blogsRouter
